+++ lib/ui/profile/ProfileInfoPage.dart
@@ -18,7 +18,9 @@ import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:rokwire_plugin/service/auth2.dart';
 import 'package:rokwire_plugin/service/content.dart';
 import 'package:rokwire_plugin/service/groups.dart';
+import 'package:rokwire_plugin/service/inbox.dart';
 import 'package:rokwire_plugin/service/localization.dart';
+import 'package:rokwire_plugin/service/network.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/social.dart';
 import 'package:rokwire_plugin/service/styles.dart';
@@ -386,20 +388,36 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
               options: (0 < userPostCount) ? [groupsSwitchTitle] : null,
               initialOptionsSelection: (0 < userPostCount) ?  [groupsSwitchTitle] : [],
               continueTitle: Localization().getStringEx("panel.settings.privacy_center.button.forget_info.title","Forget My Information"),
-              onContinue: (List<String> selectedValues, OnContinueProgressController progressController) async {
-                Analytics().logAlert(text: "Remove My Information", selection: "Yes");
-                progressController(loading: true);
-                if (selectedValues.contains(groupsSwitchTitle)){
-                  Future.wait([Groups().deleteUserData(), Social().deleteUser()]);
-                }
-                await Auth2().deleteUser();
-                progressController(loading: false);
-                Navigator.pop(context);
-              },
+              onContinue: (List<String> selectedValues, OnContinueProgressController progressController) => _deleteAccount(selectedValues.contains(groupsSwitchTitle), progressController),
               longButtonTitle: true
           );
         }
       });
+    }
+  }
+
+  void _deleteAccount(bool deleteContributions, OnContinueProgressController progressController) async {
+    Analytics().logAlert(text: "Remove My Information", selection: "Yes");
+    progressController(loading: true);
+    NetworkAuthProvider? authProvider = Auth2().networkAuthProvider; // Store token before
+    bool? result = await Auth2().deleteUser();
+    if (result == true) {
+      List<Future<bool?>> futures = [
+        Inbox().deleteUser(auth: authProvider)
+      ];
+      if (deleteContributions) {
+        futures.addAll(<Future<bool?>>[
+          Groups().deleteUserData(auth: authProvider),
+          Social().deleteUser(auth: authProvider)
+        ]);
+      }
+      await Future.wait(futures);
+      progressController(loading: false);
+      Navigator.pop(context);
+    }
+    else {
+      progressController(loading: false);
+      AppAlert.showTextMessage(context, Localization().getStringEx('panel.profile.info.delete.failed.text', 'Failed to delete app account.'));
     }
   }